# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Progi CAP API",
        "description": "API documentation for Progi CAP application modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/dossiers/periods": {
            "get": {
                "tags": [
                    "Dossier Submission"
                ],
                "summary": "Périodes de soumission actives par cycle",
                "description": "Liste les périodes de soumission actives pour un cycle donné",
                "operationId": "getDossierSubmissionPeriods",
                "parameters": [
                    {
                        "name": "cycle",
                        "in": "query",
                        "description": "Cycle cible",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Licence",
                                "Master",
                                "Ingénieur"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des périodes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "department": {
                                                        "type": "string"
                                                    },
                                                    "academic_year": {
                                                        "type": "string"
                                                    },
                                                    "start_date": {
                                                        "type": "string"
                                                    },
                                                    "end_date": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cycle invalide"
                    },
                    "404": {
                        "description": "Aucune période active"
                    }
                }
            }
        },
        "/api/dossiers/licence": {
            "post": {
                "tags": [
                    "Dossier Submission"
                ],
                "summary": "Soumettre un dossier Licence",
                "description": "Soumet un dossier pour le cycle Licence avec documents requis",
                "operationId": "submitLicenceDossier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "last_name",
                                    "first_names",
                                    "email",
                                    "birth_date",
                                    "birth_place",
                                    "birth_country",
                                    "gender",
                                    "contacts",
                                    "study_level",
                                    "academic_year_id",
                                    "department_id",
                                    "demande_da",
                                    "cv",
                                    "acte_naissance",
                                    "diplome_bac",
                                    "attestation_travail",
                                    "quittance_rectorat",
                                    "quittance_cap"
                                ],
                                "properties": {
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_names": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "birth_place": {
                                        "type": "string"
                                    },
                                    "birth_country": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ]
                                    },
                                    "contacts": {
                                        "type": "array",
                                        "items": {}
                                    },
                                    "study_level": {
                                        "type": "string"
                                    },
                                    "entry_diploma_id": {
                                        "type": "integer"
                                    },
                                    "academic_year_id": {
                                        "type": "integer"
                                    },
                                    "department_id": {
                                        "type": "integer"
                                    },
                                    "demande_da": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cv": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "acte_naissance": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "diplome_bac": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "diplome_licence": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "attestation_travail": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "quittance_rectorat": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "quittance_cap": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "attestation_depot_dossier": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soumission réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "tracking_code": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation / période inactive"
                    }
                }
            }
        },
        "/api/dossiers/master": {
            "post": {
                "tags": [
                    "Dossier Submission"
                ],
                "summary": "Soumettre un dossier Master",
                "operationId": "submitMasterDossier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "last_name",
                                    "first_names",
                                    "email",
                                    "birth_date",
                                    "birth_place",
                                    "birth_country",
                                    "gender",
                                    "contacts",
                                    "study_level",
                                    "academic_year_id",
                                    "department_id",
                                    "demande_da",
                                    "cv",
                                    "acte_naissance",
                                    "diplome_bac",
                                    "diplome_license",
                                    "attestation_travail",
                                    "quittance_rectorat",
                                    "quittance_cap"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soumission réussie"
                    },
                    "400": {
                        "description": "Erreur de validation / période inactive"
                    }
                }
            }
        },
        "/api/dossiers/ingenieur/prepa": {
            "post": {
                "tags": [
                    "Dossier Submission"
                ],
                "summary": "Soumettre un dossier Ingénieur Prépa",
                "operationId": "submitIngenieurPrepaDossier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "last_name",
                                    "first_names",
                                    "email",
                                    "birth_date",
                                    "birth_place",
                                    "birth_country",
                                    "gender",
                                    "contacts",
                                    "study_level",
                                    "academic_year_id",
                                    "department_id",
                                    "demande_da",
                                    "cv",
                                    "acte_naissance",
                                    "diplome_bac",
                                    "diplome_licence",
                                    "attestation_travail",
                                    "quittance_rectorat",
                                    "quittance_cap"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soumission réussie"
                    },
                    "400": {
                        "description": "Erreur de validation / période inactive"
                    }
                }
            }
        },
        "/api/dossiers/ingenieur/specialite": {
            "post": {
                "tags": [
                    "Dossier Submission"
                ],
                "summary": "Soumettre un dossier Ingénieur Spécialité",
                "operationId": "submitIngenieurSpecialiteDossier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "student_id_number",
                                    "department_id",
                                    "academic_year_id",
                                    "certificat_prepa"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soumission réussie"
                    },
                    "400": {
                        "description": "Non éligible / validation échouée"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}